backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.50.0
pandas>=2.2.0
python-dotenv==1.0.0
python-multipart==0.0.6 